Move WidgetId to widget/mod.rs, IdentityWrapper to widget/identity_wrapper.rs
This lets us delete the widget_id file; the motivation is that widgets
should be in the widget module. LensWrap is the only exception, and
maybe we should address that?
