+++ druid/src/widget/mod.rs
@@ -21,6 +21,7 @@ mod container;
 mod either;
 mod env_scope;
 mod flex;
+mod identity_wrapper;
 mod label;
 mod list;
 mod padding;
@@ -46,6 +47,7 @@ pub use container::Container;
 pub use either::Either;
 pub use env_scope::EnvScope;
 pub use flex::{Column, Flex, Row};
+pub use identity_wrapper::IdentityWrapper;
 pub use label::{Label, LabelText};
 pub use list::{List, ListIter};
 pub use padding::Padding;
@@ -65,9 +67,14 @@ pub use textbox::TextBox;
 pub use widget_ext::WidgetExt;
 
 use std::ops::{Deref, DerefMut};
+use std::sync::atomic::{AtomicU32, Ordering};
 
 use crate::kurbo::Size;
-use crate::{BoxConstraints, Env, Event, EventCtx, LayoutCtx, PaintCtx, UpdateCtx, WidgetId};
+use crate::{BoxConstraints, Env, Event, EventCtx, LayoutCtx, PaintCtx, UpdateCtx};
+
+/// A unique identifier for a single widget.
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct WidgetId(u32);
 
 /// The trait implemented by all widgets.
 ///
@@ -176,6 +183,19 @@ pub trait Widget<T> {
     /// `IdentityWrapper`. Widgets should not implement this on their own.
     fn id(&self) -> Option<WidgetId> {
         None
+    }
+}
+
+//TODO: we implement this in like 5 different places now :/
+static WIDGET_ID_COUNTER: AtomicU32 = AtomicU32::new(1);
+
+impl WidgetId {
+    /// Allocate a new, unique widget id.
+    ///
+    /// Do note that if we create 4 billion widgets there may be a collision.
+    pub(crate) fn next() -> WidgetId {
+        let id = WIDGET_ID_COUNTER.fetch_add(1, Ordering::Relaxed);
+        WidgetId(id)
     }
 }
